checking in further progress
