+++ watertap/unit_models/reverse_osmosis_0D.py
@@ -24,31 +24,18 @@ from pyomo.environ import (
     check_optimal_termination,
 )
 
-# Import IDAES cores
-from idaes.core import ControlVolume0DBlock, declare_process_block_class
-from idaes.core.util.exceptions import ConfigurationError
-from idaes.core.solvers import get_solver
+from idaes.core import declare_process_block_class
 from idaes.core.util.misc import add_object_reference
-import idaes.core.util.scaling as iscale
 from watertap.core import MembraneChannel0D
-from watertap.core.util.initialization import check_solve, check_dof
-from watertap.unit_models._reverse_osmosis_base import (
-    ConcentrationPolarizationType,
-    MassTransferCoefficient,
-    PressureChangeType,
-    _ReverseOsmosisBaseData,
-)
+from watertap.unit_models.reverse_osmosis_base import ReverseOsmosisBaseData
 import idaes.logger as idaeslog
 
 
 __author__ = "Tim Bartholomew, Adam Atia"
 
-# Set up logger
-_log = idaeslog.getLogger(__name__)
-
 
 @declare_process_block_class("ReverseOsmosis0D")
-class ReverseOsmosisData(_ReverseOsmosisBaseData):
+class ReverseOsmosisData(ReverseOsmosisBaseData):
     """
     Standard RO Unit Model Class:
     - zero dimensional model
@@ -56,7 +43,7 @@ class ReverseOsmosisData(_ReverseOsmosisBaseData):
     - single liquid phase only
     """
 
-    CONFIG = _ReverseOsmosisBaseData.CONFIG()
+    CONFIG = ReverseOsmosisBaseData.CONFIG()
 
     def _add_membrane_channel(self):
         # Build membrane channel control volume
@@ -68,155 +55,3 @@ class ReverseOsmosisData(_ReverseOsmosisBaseData):
                 "property_package_args": self.config.property_package_args,
             }
         )
-
-    def initialize_build(
-        blk,
-        initialize_guess=None,
-        state_args=None,
-        outlvl=idaeslog.NOTSET,
-        solver=None,
-        optarg=None,
-        fail_on_warning=False,
-        ignore_dof=False,
-    ):
-        """
-        General wrapper for RO initialization routines
-
-        Keyword Arguments:
-
-            initialize_guess : a dict of guesses for solvent_recovery, solute_recovery,
-                               and cp_modulus. These guesses offset the initial values
-                               for the retentate, permeate, and membrane interface
-                               state blocks from the inlet feed
-                               (default =
-                               {'deltaP': -1e4,
-                               'solvent_recovery': 0.5,
-                               'solute_recovery': 0.01,
-                               'cp_modulus': 1.1})
-            state_args : a dict of arguments to be passed to the property
-                         package(s) to provide an initial state for the inlet
-                         feed side state block (see documentation of the specific
-                         property package) (default = None).
-            outlvl : sets output level of initialization routine
-            optarg : solver options dictionary object (default=None)
-            solver : solver object or string indicating which solver to use during
-                     initialization, if None provided the default solver will be used
-                     (default = None)
-            fail_on_warning : boolean argument to fail or only produce  warning upon unsuccessful solve (default=False)
-            ignore_dof : boolean argument to ignore when DOF != 0 (default=False)
-        Returns:
-            None
-        """
-
-        init_log = idaeslog.getInitLogger(blk.name, outlvl, tag="unit")
-        solve_log = idaeslog.getSolveLogger(blk.name, outlvl, tag="unit")
-        # Set solver and options
-        opt = get_solver(solver, optarg)
-
-        # ---------------------------------------------------------------------
-        # Extract initial state of inlet feed
-        source = blk.feed_side.properties_in[blk.flowsheet().config.time.first()]
-        state_args = blk._get_state_args(
-            source, blk.mixed_permeate[0], initialize_guess, state_args
-        )
-
-        # Initialize feed inlet state block
-        flags_feed_side = blk.feed_side.properties_in.initialize(
-            outlvl=outlvl,
-            optarg=optarg,
-            solver=solver,
-            state_args=state_args["feed_side"],
-            hold_state=True,
-        )
-
-        init_log.info("Initialization Step 1 Complete.")
-        if not ignore_dof:
-            check_dof(blk, fail_flag=fail_on_warning, logger=init_log)
-        # ---------------------------------------------------------------------
-        # Initialize other state blocks
-        # base properties on inlet state block
-
-        blk.feed_side.properties_out.initialize(
-            outlvl=outlvl,
-            optarg=optarg,
-            solver=solver,
-            state_args=state_args["retentate"],
-        )
-        blk.feed_side.properties_interface.initialize(
-            outlvl=outlvl,
-            optarg=optarg,
-            solver=solver,
-            state_args=state_args["interface"],
-        )
-        blk.mixed_permeate.initialize(
-            outlvl=outlvl,
-            optarg=optarg,
-            solver=solver,
-            state_args=state_args["permeate"],
-        )
-        blk.permeate_side.initialize(
-            outlvl=outlvl,
-            optarg=optarg,
-            solver=solver,
-            state_args=state_args["permeate"],
-        )
-        init_log.info("Initialization Step 2 Complete.")
-
-        # ---------------------------------------------------------------------
-        # Solve unit
-        with idaeslog.solver_log(solve_log, idaeslog.DEBUG) as slc:
-            res = opt.solve(blk, tee=slc.tee)
-            # occasionally it might be worth retrying a solve
-            if not check_optimal_termination(res):
-                init_log.warning(
-                    "Trouble solving ReverseOsmosis0D unit model, trying one more time"
-                )
-                res = opt.solve(blk, tee=slc.tee)
-        check_solve(
-            res,
-            checkpoint="Initialization Step 3",
-            logger=init_log,
-            fail_flag=fail_on_warning,
-        )
-        # ---------------------------------------------------------------------
-        # Release Inlet state
-        blk.feed_side.release_state(flags_feed_side, outlvl)
-        init_log.info("Initialization Complete: {}".format(idaeslog.condition(res)))
-
-    def calculate_scaling_factors(self):
-        # setting scaling factors for variables
-        # will not override if the user does provide the scaling factor
-        if iscale.get_scaling_factor(self.dens_solvent) is None:
-            sf = iscale.get_scaling_factor(
-                self.feed_side.properties_in[0].dens_mass_phase["Liq"]
-            )
-            iscale.set_scaling_factor(self.dens_solvent, sf)
-
-        super().calculate_scaling_factors()
-
-        for (t, p, j), v in self.mass_transfer_phase_comp.items():
-            sf = iscale.get_scaling_factor(
-                self.feed_side.properties_in[t].get_material_flow_terms(p, j)
-            )
-            if iscale.get_scaling_factor(v) is None:
-                iscale.set_scaling_factor(v, sf)
-            v = self.feed_side.mass_transfer_term[t, p, j]
-            if iscale.get_scaling_factor(v) is None:
-                iscale.set_scaling_factor(v, sf)
-
-        if hasattr(self, "area_cross"):
-            if iscale.get_scaling_factor(self.area_cross) is None:
-                iscale.set_scaling_factor(self.area_cross, 100)
-
-        if hasattr(self, "length"):
-            if iscale.get_scaling_factor(self.length) is None:
-                iscale.set_scaling_factor(self.length, 1)
-
-        if hasattr(self, "width"):
-            if iscale.get_scaling_factor(self.width) is None:
-                iscale.set_scaling_factor(self.width, 1)
-
-        if hasattr(self, "dP_dx"):
-            for v in self.dP_dx.values():
-                if iscale.get_scaling_factor(v) is None:
-                    iscale.set_scaling_factor(v, 1e-4)
