+++ watertap/unit_models/reverse_osmosis_1D.py
@@ -24,25 +24,15 @@ from pyomo.common.config import ConfigValue, In
 
 # Import IDAES cores
 from idaes.core import (
-    ControlVolume1DBlock,
     declare_process_block_class,
-    MomentumBalanceType,
     useDefault,
 )
-from idaes.core.base.control_volume1d import DistributedVars
 from idaes.core.util.misc import add_object_reference
-from idaes.core.solvers import get_solver
-from idaes.core.util import scaling as iscale
-from idaes.core.util.initialization import solve_indexed_blocks
-from watertap.core.util.initialization import check_solve, check_dof
-from watertap.unit_models._reverse_osmosis_base import (
-    ConcentrationPolarizationType,
-    MassTransferCoefficient,
-    PressureChangeType,
-    _ReverseOsmosisBaseData,
-)
 import idaes.logger as idaeslog
 
+from watertap.core import MembraneChannel1D
+from watertap.core.membrane_channel1d import CONFIG_Template
+from watertap.unit_models.reverse_osmosis_base import ReverseOsmosisBaseData
 
 __author__ = "Adam Atia"
 
@@ -51,67 +41,10 @@ _log = idaeslog.getLogger(__name__)
 
 
 @declare_process_block_class("ReverseOsmosis1D")
-class ReverseOsmosis1DData(_ReverseOsmosisBaseData):
+class ReverseOsmosis1DData(ReverseOsmosisBaseData):
     """Standard 1D Reverse Osmosis Unit Model Class."""
 
-    CONFIG = _ReverseOsmosisBaseData.CONFIG()
-
-    CONFIG.declare(
-        "area_definition",
-        ConfigValue(
-            default=DistributedVars.uniform,
-            domain=In(DistributedVars),
-            description="Argument for defining form of area variable",
-            doc="""Argument defining whether area variable should be spatially
-    variant or not. **default** - DistributedVars.uniform.
-    **Valid values:** {
-    DistributedVars.uniform - area does not vary across spatial domain,
-    DistributedVars.variant - area can vary over the domain and is indexed
-    by time and space.}""",
-        ),
-    )
-
-    CONFIG.declare(
-        "transformation_method",
-        ConfigValue(
-            default=useDefault,
-            description="Discretization method to use for DAE transformation",
-            doc="""Discretization method to use for DAE transformation. See Pyomo
-    documentation for supported transformations.""",
-        ),
-    )
-
-    CONFIG.declare(
-        "transformation_scheme",
-        ConfigValue(
-            default=useDefault,
-            description="Discretization scheme to use for DAE transformation",
-            doc="""Discretization scheme to use when transforming domain. See
-    Pyomo documentation for supported schemes.""",
-        ),
-    )
-
-    CONFIG.declare(
-        "finite_elements",
-        ConfigValue(
-            default=20,
-            domain=int,
-            description="Number of finite elements in length domain",
-            doc="""Number of finite elements to use when discretizing length 
-            domain (default=20)""",
-        ),
-    )
-
-    CONFIG.declare(
-        "collocation_points",
-        ConfigValue(
-            default=5,
-            domain=int,
-            description="Number of collocation points per finite element",
-            doc="""Number of collocation points to use per finite element when
-            discretizing length domain (default=5)""",
-        ),
-    )
+    CONFIG = CONFIG_Template.CONFIG()
 
     def _process_config(self):
         if self.config.transformation_method is useDefault:
@@ -160,175 +93,11 @@ class ReverseOsmosis1DData(_ReverseOsmosisBaseData):
         Returns:
             None
         """
-        # Call UnitModel.build to setup dynamics
-        super().build()
-
         # Check configuration errors
         self._process_config()
 
+        # Call UnitModel.build to setup dynamics
+        super().build()
+
         # ==========================================================================
         """ Add references to control volume geometry."""
-        add_object_reference(self, "length", feed_side.length)
-        add_object_reference(self, "area_cross", feed_side.area)
-
-
-    def initialize_build(
-        blk,
-        initialize_guess=None,
-        state_args=None,
-        outlvl=idaeslog.NOTSET,
-        solver=None,
-        optarg=None,
-        fail_on_warning=False,
-        ignore_dof=False,
-    ):
-        """
-        Initialization routine for 1D-RO unit.
-
-        Keyword Arguments:
-            initialize_guess : a dict of guesses for solvent_recovery, solute_recovery,
-                               and cp_modulus. These guesses offset the initial values
-                               for the retentate, permeate, and membrane interface
-                               state blocks from the inlet feed
-                               (default =
-                               {'deltaP': -1e4,
-                               'solvent_recovery': 0.5,
-                               'solute_recovery': 0.01,
-                               'cp_modulus': 1.1})
-            state_args : a dict of arguments to be passed to the property
-                         package(s) to provide an initial state for the inlet
-                         feed side state block (see documentation of the specific
-                         property package) (default = None).
-            outlvl : sets output level of initialization routine
-            solver : str indicating which solver to use during
-                     initialization (default = None, use default solver)
-            optarg : solver options dictionary object (default=None, use default solver options)
-            fail_on_warning : boolean argument to fail or only produce  warning upon unsuccessful solve (default=False)
-            ignore_dof : boolean argument to ignore when DOF != 0 (default=False)
-        Returns:
-            None
-
-        """
-
-        init_log = idaeslog.getInitLogger(blk.name, outlvl, tag="unit")
-        solve_log = idaeslog.getSolveLogger(blk.name, outlvl, tag="unit")
-
-        # Create solver
-        opt = get_solver(solver, optarg)
-
-        source = blk.feed_side.properties[
-            blk.flowsheet().config.time.first(), blk.first_element
-        ]
-        state_args = blk._get_state_args(
-            source, blk.mixed_permeate[0], initialize_guess, state_args
-        )
-
-        # ---------------------------------------------------------------------
-        # Step 1: Initialize feed_side, permeate_side, and mixed_permeate blocks
-        flags_feed_side = blk.feed_side.initialize(
-            outlvl=outlvl,
-            optarg=optarg,
-            solver=solver,
-            state_args=state_args["feed_side"],
-            hold_state=True,
-        )
-
-        init_log.info("Initialization Step 1 Complete")
-        if not ignore_dof:
-            check_dof(blk, fail_flag=fail_on_warning, logger=init_log)
-        # ---------------------------------------------------------------------
-        # Initialize other state blocks
-        # base properties on inlet state block
-
-        flag_feed_side_properties_interface = (
-            blk.feed_side.properties_interface.initialize(
-                outlvl=outlvl,
-                optarg=optarg,
-                solver=solver,
-                state_args=state_args["interface"],
-            )
-        )
-        flags_permeate_side = blk.permeate_side.initialize(
-            outlvl=outlvl,
-            optarg=optarg,
-            solver=solver,
-            state_args=state_args["permeate"],
-        )
-        flags_mixed_permeate = blk.mixed_permeate.initialize(
-            outlvl=outlvl,
-            optarg=optarg,
-            solver=solver,
-            state_args=state_args["permeate"],
-        )
-        init_log.info("Initialization Step 2 Complete.")
-
-        # ---------------------------------------------------------------------
-        # Solve unit
-        with idaeslog.solver_log(solve_log, idaeslog.DEBUG) as slc:
-            res = opt.solve(blk, tee=slc.tee)
-            # occasionally it might be worth retrying a solve
-            if not check_optimal_termination(res):
-                init_log.warning(
-                    "Trouble solving ReverseOsmosis1D unit model, trying one more time"
-                )
-                res = opt.solve(blk, tee=slc.tee)
-        check_solve(
-            res,
-            logger=init_log,
-            fail_flag=fail_on_warning,
-            checkpoint="Initialization Step 3",
-        )
-        # ---------------------------------------------------------------------
-        # Release Inlet state
-        blk.feed_side.release_state(flags_feed_side, outlvl)
-        init_log.info("Initialization Complete: {}".format(idaeslog.condition(res)))
-
-    def calculate_scaling_factors(self):
-        if iscale.get_scaling_factor(self.dens_solvent) is None:
-            sf = iscale.get_scaling_factor(
-                self.feed_side.properties[0, 0].dens_mass_phase["Liq"]
-            )
-            iscale.set_scaling_factor(self.dens_solvent, sf)
-
-        super().calculate_scaling_factors()
-
-        # these variables should have user input, if not there will be a warning
-        if iscale.get_scaling_factor(self.width) is None:
-            sf = iscale.get_scaling_factor(self.width, default=1, warning=True)
-            iscale.set_scaling_factor(self.width, sf)
-
-        if iscale.get_scaling_factor(self.length) is None:
-            sf = iscale.get_scaling_factor(self.length, default=10, warning=True)
-            iscale.set_scaling_factor(self.length, sf)
-
-        # setting scaling factors for variables
-
-        # will not override if the user provides the scaling factor
-        ## default of 1 set by ControlVolume1D
-        if iscale.get_scaling_factor(self.area_cross) == 1:
-            iscale.set_scaling_factor(self.area_cross, 100)
-
-        for (t, x, p, j), v in self.mass_transfer_phase_comp.items():
-            sf = (
-                iscale.get_scaling_factor(
-                    self.feed_side.properties[t, x].get_material_flow_terms(p, j)
-                )
-                / iscale.get_scaling_factor(self.feed_side.length)
-            ) * value(self.nfe)
-            if iscale.get_scaling_factor(v) is None:
-                iscale.set_scaling_factor(v, sf)
-            v = self.feed_side.mass_transfer_term[t, x, p, j]
-            if iscale.get_scaling_factor(v) is None:
-                iscale.set_scaling_factor(v, sf)
-
-        if hasattr(self, "deltaP"):
-            for v in self.deltaP.values():
-                if iscale.get_scaling_factor(v) is None:
-                    iscale.set_scaling_factor(v, 1e-4)
-
-        if hasattr(self, "dP_dx"):
-            for v in self.feed_side.pressure_dx.values():
-                iscale.set_scaling_factor(v, 1e-5)
-        else:
-            for v in self.feed_side.pressure_dx.values():
-                iscale.set_scaling_factor(v, 1e5)
